requirements-dev.txt
-i https://pypi.org/simple
alabaster==0.7.12
apipkg==1.5
appdirs==1.4.3
arpeggio==1.9.0
atomicwrites==1.2.1
attrs==18.2.0
babel==2.6.0
black==18.9b0
certifi==2018.10.15
chardet==3.0.4
click==7.0
codacy-coverage==1.3.11
colorama==0.4.0
coverage==4.5.1
docutils==0.14
execnet==1.5.0
filelock==3.0.10
flake8==3.6.0
hypothesis==3.82.1
idna==2.7
imagesize==1.1.0
incremental==17.5.0
invoke==1.2.0
isort==4.3.4
jinja2==2.10
lxml==4.2.5
markupsafe==1.1.0
mccabe==0.6.1
more-itertools==4.3.0
msgpack==0.5.6
packaging==18.0
parver==0.1.1
pluggy==0.8.0
psutil==5.4.8
py==1.7.0
pycodestyle==2.4.0
pyflakes==2.0.0
pygments==2.2.0
pyparsing==2.3.0
pytest-cov==2.6.0
pytest-flake8==1.0.2
pytest-forked==0.2
pytest-sugar==0.9.2
pytest-xdist==1.24.1
pytest==3.9.3
python-rapidjson==0.6.3
pytoml==0.1.20
pytz==2018.7
pyyaml==4.2b4
requests==2.20.1
six==1.11.0
snowballstemmer==1.2.1
sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
termcolor==1.1.0
toml==0.10.0
tomlkit==0.5.2
towncrier==18.6.0
tox==3.5.3
ujson==1.35
urllib3==1.24.1
virtualenv==16.1.0
vprof==0.37.6
